requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.0
install==1.3.5
kombu==5.3.5
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.2
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
